% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/division.R
\name{div}
\alias{div}
\title{Division entre dos numeros}
\usage{
div(a, b)
}
\arguments{
\item{a}{Primer numero}

\item{b}{Segundo numero (Distinto de 0)}
}
\value{
La division entre a y b
}
\description{
Esta funcion devuelve la division entre dos numeros.
}
\examples{
div(5,0.5)
div(3,2)
div(1,0.0004)

}
\seealso{
\link{mult}
}
