% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicacion.R
\name{mult}
\alias{mult}
\title{Multiplicacion entre dos numeros}
\usage{
mult(a, b)
}
\arguments{
\item{a}{Primer numero}

\item{b}{Segundo numero}
}
\value{
La multiplicacion entre a y b
}
\description{
Esta funcion devuelve la multiplicacion entre dos numeros.
}
\examples{
mult(2,3.5)

}
\seealso{
\link{div}
}
