% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resta.R
\name{resta}
\alias{resta}
\title{Resta entre dos numeros}
\usage{
resta(a, b)
}
\arguments{
\item{a}{Primer numero}

\item{b}{Segundo numero}
}
\value{
La resta entre a y b
}
\description{
Esta funcion devuelve la resta entre dos numeros.
}
\examples{
resta(5.2,6)

}
\seealso{
\link{suma}
}
