% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma.R
\name{suma}
\alias{suma}
\title{Suma de dos numeros}
\usage{
suma(a, b)
}
\arguments{
\item{a}{Primer numero}

\item{b}{Segundo numero}
}
\value{
La suma entre a y b
}
\description{
Esta funcion devuelve la suma entre dos numeros.
}
\examples{
suma(5,3)

}
\seealso{
\link{resta}
}
