% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div.R
\name{div}
\alias{div}
\title{División entre dos números}
\usage{
div(a, b)
}
\arguments{
\item{a}{Primer número}

\item{b}{Segundo número (Distinto de 0).}
}
\value{
La división entre a y b
}
\description{
Esta función devuelve la división entre dos números. El segundo número
no puede ser negativo.
}
\examples{
div(5, 0.5)

}
\seealso{
\link{mult}
}
