% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult.R
\name{mult}
\alias{mult}
\title{Multiplicación entre dos números}
\usage{
mult(a, b)
}
\arguments{
\item{a}{Primer número}

\item{b}{Segundo número}
}
\value{
La multiplicación entre a y b
}
\description{
Esta función devuelve la multiplicación entre dos números.
}
\examples{
mult(2, 3.5)

}
\seealso{
\link{div}
}
