% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma.R
\name{suma}
\alias{suma}
\title{Suma entre dos números}
\usage{
suma(a, b)
}
\arguments{
\item{a}{Primer número}

\item{b}{Segundo número}
}
\value{
La suma entre a y b
}
\description{
Esta función devuelve la suma entre dos números.
}
\examples{
suma(5, 3)

}
\seealso{
\link{resta}
}
